backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = ""
authors = ["nyvrae <mex99anue@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
sqlalchemy = "^2.0.43"
playwright = "^1.54.0"
uvicorn = "^0.32.1"
pydantic = {extras = ["email"], version = "^2.11.7"}
python-dotenv = "^1.1.1"
pyjwt = {extras = ["crypto"], version = "^2.10.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
celery = "^5.5.3"
redis = "^6.4.0"
flower = "^2.0.1"
gevent = "^25.8.2"

[tool.poetry.group.dev.dependencies]
beautifulsoup4 = "^4.13.4"
asyncio = "^4.0.0"
pytest = "^8.4.2"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
